+++ shared/common/src/components/Auth/Login/LoginForm/LoginButtons/style.ts
@@ -1,0 +1,22 @@
+import styled from '@emotion/styled'
+
+export const LoginButtonWrapper = styled.div`
+  width: 100%;
+  height: 6.625rem;
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+`
+
+export const LoginButton = styled.button<{ isAble: boolean }>`
+  background-color: ${({ theme, isAble }) =>
+    isAble ? theme.color.main : theme.color.gray['700']};
+  color: ${({ theme, isAble }) =>
+    isAble ? theme.color.white : theme.color.gray['400']};
+  ${({ theme }) => theme.typo.text_lg.semibold};
+  width: 24rem;
+  height: 3.25rem;
+  border-radius: 0.5rem;
+  border: 0;
+  cursor: pointer;
+`
