+++ shared/common/src/components/Auth/Login/LoginForm/LoginInput/style.ts
@@ -1,0 +1,22 @@
+import styled from '@emotion/styled'
+
+export const InputWrapper = styled.div`
+  width: 100%;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+`
+
+export const InputContainer = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.5rem;
+
+  span {
+    width: 100%;
+    margin-left: 3.2rem;
+    margin-top: -1rem;
+  }
+`
