+++ shared/common/src/components/Auth/Login/LoginForm/index.tsx
@@ -1,0 +1,13 @@
+import LoginInput from "./LoginInput"
+import LoginButtons from "./LoginButtons"
+
+const LoginForm = () => {
+  return (
+    <>
+      <LoginInput />
+      <LoginButtons />
+    </>
+  )
+}
+
+export default LoginForm
