+++ shared/common/src/components/Auth/Login/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginForm } from './LoginForm'
+export { default as LoginBottom } from './LoginBottom'
