+++ shared/common/src/components/Auth/index.ts
@@ -1,0 +1,1 @@
+export * from './Login'
