+++ shared/common/src/components/index.ts
@@ -3,6 +3,7 @@ export { default as CertificateItem } from './CertificateItem'
 export * from './ClubListContents'
 export { default as CommonCheckBox } from './CommonCheckBox'
 export * from './CompanyListContents'
+export * from './Auth'
 export { default as CompleteLectureItem } from './CompleteLectureItem'
 export { default as FAQAnswerItem } from './FAQAnswerItem'
 export { default as FAQItem } from './FAQItem'
